// src/types/upgradeCustomer.ts
export interface IUserUpgradeType {
  user_photo: string;
  first_name: string;
  last_name: string;
  email: string;
  mobile: string;
  organisation: string;
  designation: string;
  experience: string;
  award: string[];
  country: string;
  zip_code: string;
  city: string;
  address: string;
  about: string;
  website: string;
  fb_url: string;
  twitter_url: string;
  insta_url: string;
  youtube_url: string;
}